Made *Mobile* class independent from *App* class. Since both class are kind on the same level with both particular layout, I think it is best to have both separated.
Made the *sort_dictionary_to_list* method a global function.

Created a new class *BrowseAndSubmit* to manage browsing for new files, displaying the results in a QLabel and activating/deactivating the *submit button*. An instance of the class is created in both *Mobile* and *App*. IMO it is cleaner this way then using signal and alpha function for two reasons: First, it reduces the amount of coding necessary, since the code is reused in two classes. Second, signal and alpha function can be trickly later on when one whish to optimize the memory print taken by the program.